more cleaning up in the RNA code...
